InsightGUIDE/backend-api/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
gunicorn>=21.0.0,<22.0.0
mistralai>=1.0.0,<2.0.0
openai>=1.12.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0.1,<7.0.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6,<1.0.0
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
httpx>=0.26.0,<1.0.0